+++ src/dflow/config.py
@@ -1,0 +1,8 @@
+config = {
+    "host": "https://127.0.0.1:2746",
+    "token": None,
+    "k8s_config_file": None,
+    "k8s_api_server": None,
+    "private_key_host_path": "/home/docker/.ssh",
+    "save_path_as_parameter": True
+}